Move add to schedule to showTalk component
