+++ src/services/api-service.js
@@ -9,9 +9,6 @@ const get = (onSuccess, onError, apiURL) => {
     }
   }).catch(error => onError(error));
 };
-const get1 = (apiURL) => {
-  fetch(apiURL, {method: 'GET'}).then(result => result.json())
-};
 
 const post = ( onSuccess, onError, data) => {
   fetch('http://localhost:3001/schedules', {
@@ -20,7 +17,15 @@ const post = ( onSuccess, onError, data) => {
       'Content-Type': 'application/json', 
     },
     body: JSON.stringify(data),
-  })
+  }).then(result => result.json())
+  .then((result) => {
+    if (result.error) {
+      onError();
+    }
+    else {
+      onSuccess(result);
+    }
+  }).catch(error => onError(error));
 };
 
 const remove = (onSuccess, onError, apiURL) => {
